Code cleanup, compatibility updates, and logging improvements.
This commit includes several minor changes across different modules, encompassing exception handling, print statement updates for Python 3 compatibility, and updates to library imports for broader compatibility. It also improves the structure of logging messages and ensures data is handled in the correct format.

Changes
STM32Loader/stm32_check_cal.py: Replaced a bare except clause with except Exception for better exception handling and added parenthesis to the print statement for Python 3 compatibility.
flightcode/dflog/dataFlashMAVLink-to-artoo.py: Added import sys, provided a default value to link_dir in the RecentLogLinker class initializer, and updated print statements for Python 3 compatibility. Also handles decoding stdout/stderr from subprocess calls.
flightcode/dflog/loadLog.py: Updated print statements for Python 3 compatibility.
flightcode/old/app_streamer.py: Changed ConfigParser import to configparser as ConfigParser and cast log_interval_us to int.
flightcode/old/logd.py: Changed cPickle import to pickle, added protocol=pickle.HIGHEST_PROTOCOL to pickle.dumps to handle larger data structures, corrected type annotations, ensured raw packets are properly encoded before logging, and decoded strings from bytes objects when logging.
flightcode/old/mavproxy_solo.py: Updated print statements for Python 3 compatibility.
flightcode/old/pixrc.py: Changed ConfigParser import to configparser as ConfigParser.
flightcode/old/rssi_send.py: Updated print statement for Python 3 compatibility and ensured rssi is an integer before printing.
flightcode/old/slip.py: Added handling for bytes types in _as_char method to properly decode incoming data. Added encoding for latin1 on the send path.
flightcode/old/stm32.py: Changed ConfigParser import to configparser as ConfigParser and updated print statements for Python 3 compatibility.
flightcode/old/stm32_sim.py: Updated print statements for Python 3 compatibility, and decode struct.pack output with latin1
flightcode/python/SoloLED.py: Updated print statements for Python 3 compatibility and used bytearray literals and extend method to construct the message.
flightcode/python/app_server.py: Updated print statements for Python 3 compatibility and use bytearray literals instead of strings for pkt
flightcode/python/attic/checkPixBaud.py: Changed ConfigParser import to configparser as ConfigParser and updated print statements for Python 3 compatibility.
flightcode/python/gpio.py: Updated print statements for Python 3 compatibility.
flightcode/python/led_control.py: Changed ledChoices assignment from choices= to a static assignment and updated print statements for Python 3 compatibility.
flightcode/python/pixhawk.py: Changed ConfigParser import to configparser as ConfigParser, updated print statements for Python 3 compatibility and open file with 'w' instead of 'wb'.
flightcode/python/slip.py: Added handling for bytes types in _as_char method to properly decode incoming data and encodes the output with latin1.
flightcode/python/stick-cal.py: Uses sys.maxsize instead of sys.maxint, decode byte strings from the socket on read, and convert the output to a bytestring when writing.
flightcode/stm32/app_connected_msg.py: Adds sys.exit(1) to the if and else statements to terminate the program and updated print statements for Python 3 compatibility.
flightcode/stm32/btn_client.py: Updated print statements for Python 3 compatibility with %s format.
flightcode/stm32/config_sweep_time.py: Adds sys.exit(1) to the else statement to terminate the program and updated print statements for Python 3 compatibility with %d format.
flightcode/stm32/input_report_client.py: Updated print statements for Python 3 compatibility.
flightcode/stm32/lockout_msg.py: Adds sys.exit(1) to the if and else statements to terminate the program and updated print statements for Python 3 compatibility.
flightcode/stm32/param_stored_vals_msg.py: Adds sys.exit(1) to the if and else statements to terminate the program and updated print statements for Python 3 compatibility.
gimbal/loadGimbal.py: Decode output from subprocess for Python 3 compatibility.
net/usr/bin/clock.py: Uses floor division // instead of / to avoid returning floats, and updated print statements for Python 3 compatibility.
net/usr/bin/getmaclocal.py: Updated print statements for Python 3 compatibility.
net/usr/bin/rc_cli.py: Send encoded string to socket for Python 3 compatibility and updated print statements for Python 3 compatibility.
pair/pair_button.py: Open the file in byte mode to avoid encode/decode problems.
px_uploader/loadPixhawk.py: Updated print statements for Python 3 compatibility.
Impact
Improved Python 3 compatibility by updating print statements and import statements.
Enhanced exception handling in stm32_check_cal.py.
Improved logging and data handling across several modules.
Potential improved interoperability across different system configurations.

+++ flightcode/old/app_streamer.py
@@ -3,7 +3,7 @@
 # This uses the above '#!' instead of '#!/usr/bin/env python' so that the
 # busybox 'pidof' can find this process by name.
 
-import ConfigParser
+import configparser as ConfigParser
 import clock
 import logging
 import logging.config
@@ -58,7 +58,7 @@ byte_count = 0
 now_us = clock.gettime_us(clock.CLOCK_MONOTONIC)
 
 log_interval_s = 10
-log_interval_us = long(log_interval_s * 1000000)
+log_interval_us = int(log_interval_s * 1000000)
 log_time_us = now_us + log_interval_us
 
 app_time_us = now_us
